modularizacion del css y los componentes
