+++ src/components/BotonEspecial.css
@@ -1,0 +1,8 @@
+.boton__sprite--shiny {
+    transform: rotate(0);
+    text-transform: uppercase;
+    background: rgb(255,175,69);
+    background: radial-gradient(circle, rgba(255,175,69,1) 0%, rgba(249,214,137,1) 100%);
+    width: 90px;
+    text-shadow: white -1px 1px;
+}
