+++ src/components/BotonGrande.jsx
@@ -1,3 +1,5 @@
+import './BotonGrande.css';
+
 function BotonGrande ({handler}) {
     return (
         <button className="boton--grande" onClick={handler}></button>
